==> TrifoPok/6_TrifoPok_Calc6_StatsDas5Trifo.Rmd <==
---
title: "Trifoliates IPok "
author: "CT"
date: "31.11.2017"
output: 
  word_document: default
  pdf_document: default
  html_document: default
---

Purpose:

-Analysis of trifoliates expansion for Poukawa
- Calculate stat for the days to reach 5 trifoliates 
- Run ANOVA for variable "Number of days to reach 5 Trifoliates " 
 
```{r loadLibraries}
library(tidyr)
library (dplyr)
library (ggplot2)
library(agricolae)
library (lubridate)
library(mosaic)
library(lattice)
library(knitr)
```

Read document generated in script "3" 5Trifo_rate

```{r}
setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\TrifoPok")
getwd()
#create file
df_trifo <- read.table("df_DAS_5_TrifoPok.txt",header=TRUE)
head(df_trifo)


```
Define formats and calculate the number of trifoliates considering the sub samples (S1-S10)

```{r loadFile}


# define factors for the ones that "look like" values
df_trifo$Plot <- as.factor(df_trifo$Plot)
df_trifo$Block <- as.factor(df_trifo$Block)
df_trifo$Cultivar <- as.factor(df_trifo$Cultivar)
df_trifo$SowTreat <- as.factor(df_trifo$SowTreat)

 df_trifo <- df_trifo %>%
  mutate(SowingDate = ymd(SowingDate)) 

str(df_trifo)



```




##Run stats for the number of days to reach 5 ttrifoliates (DAS5T) - general use Sow Treat as Factor 

```{r}
 
  file <- df_trifo


file.subs <- file 

head(file.subs)
summary(file.subs)

#ANOVA

my.anova <- aov(DAS5T ~ SowTreat + Block, data = file.subs)
#my.anova <- aov(DAS5T ~ Cultivar + Block, data = file.subs)

summary(my.anova)

```


```{r}
(LSD.test(my.anova, c("SowTreat"), alpha= 0.05, p.adj="none"))
#(LSD.test(my.anova, c("Cultivar"), alpha= 0.05, p.adj="none"))

```



## Create index to filter for loop ANOVA

```{r}
y <- c("S1","S2","S3","S4","S5","S6","S7","S8")

sdIndex <- unique(df_trifo$SowTreat)

sdIndex[order(match(sdIndex,y))]

length(sdIndex)
sdIndex
```
## Then loop anova for each individual sowing date 
```{r}
for (i in 1:length(sdIndex)) {
  
 #print(i)
 print(paste0(sdIndex[i],"--------------------------------"))
  
  file.subset <- df_trifo %>%
  filter(SowTreat==sdIndex[i])
  
 
head(file.subset, 50)
   summary(file.subset)
my.anova <- aov(DAS5T ~ Cultivar + Block, data = file.subset)
print(summary(my.anova))

if ((summary(my.anova)[[1]][["Pr(>F)"]][1])<0.05) {
  print("***Sig Cultivar***") 
  print((LSD.test(my.anova, c("Cultivar"), alpha= 0.05, p.adj="none"))) } 


if ((summary(my.anova)[[1]][["Pr(>F)"]][1])>0.05) {
  print("***Sig Cultivar***") 
  print((LSD.test(my.anova, c("Cultivar"), alpha= 0.05, p.adj="none"))) } 

# make mean separation if significant

(LSD.test(my.anova, c("Cultivar"), alpha= 0.05, p.adj="none"))
  
}

```






End of this script Calc 2. 
Move to script 3 to calculate rate of leaf appearance
